pass through window code cleaned up
